# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreditOut": {
        "description": "One source a garden's numbers rest on (doc 106).\n\nA credit is not a caption: CC-BY-4.0, dl-de/by-2-0 and the Copernicus terms\nall require the named credit, and a height shown without it is a height\nused outside its licence.",
        "properties": {
          "about": {
            "title": "About",
            "type": "string"
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "licence": {
            "title": "Licence",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "about",
          "name",
          "licence",
          "attribution"
        ],
        "title": "CreditOut",
        "type": "object"
      },
...
      "GardenCreate": {
        "description": "Creating a garden. Latitude and longitude are range-checked here, before\nthey reach solar code that would happily compute a sun path for latitude 500.",
        "properties": {
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "title": "GardenCreate",
        "type": "object"
      },
      "GardenCreated": {
        "description": "The token is returned; the numeric id deliberately is not.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "share_token": {
            "title": "Share Token",
            "type": "string"
          }
        },
        "required": [
          "share_token",
          "name"
        ],
        "title": "GardenCreated",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LightMap": {
        "description": "Mean daily sun hours per cell, row-major from the south-west corner.\n\n`stale` is the honest half. The map is expensive enough to store, so it can\nbe out of date \u2014 and a map that is quietly out of date is worse than one\nthat says so. It is computed by comparing a signature of the shading inputs,\nnot by remembering which actions ought to have invalidated it.",
        "properties": {
          "cell_m": {
            "title": "Cell M",
            "type": "number"
          },
          "cols": {
            "title": "Cols",
            "type": "integer"
          },
          "computed_at": {
            "title": "Computed At",
            "type": "string"
          },
          "hours": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Hours",
            "type": "array"
          },
          "max_hours": {
            "title": "Max Hours",
            "type": "number"
          },
          "min_x": {
            "title": "Min X",
            "type": "number"
          },
          "min_y": {
            "title": "Min Y",
            "type": "number"
          },
          "misplaced": {
            "items": {
              "$ref": "#/components/schemas/MisplacedOut"
            },
            "title": "Misplaced",
            "type": "array"
          },
          "morning": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Morning",
            "type": "array"
          },
          "roof": {
            "items": {
              "type": "boolean"
            },
            "title": "Roof",
            "type": "array"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "cell_m",
          "min_x",
          "min_y",
          "cols",
          "rows",
          "hours",
          "roof",
          "max_hours",
          "computed_at",
          "stale",
          "morning",
          "misplaced"
        ],
        "title": "LightMap",
        "type": "object"
      },
...
      "MisplacedOut": {
        "description": "A planting standing in light it did not ask for.\n\nA warning, never a refusal: a gardener may know something the model does\nnot \u2014 a cultivar bred for shade, a wall that throws light back, or simply\nthat they want it there.",
        "properties": {
          "bed_id": {
            "title": "Bed Id",
            "type": "integer"
          },
          "gets": {
            "title": "Gets",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planting_id": {
            "title": "Planting Id",
            "type": "integer"
          },
          "problem": {
            "title": "Problem",
            "type": "string"
          },
          "sun_hours": {
            "title": "Sun Hours",
            "type": "number"
          },
          "taxon_id": {
            "title": "Taxon Id",
            "type": "integer"
          },
          "wants": {
            "title": "Wants",
            "type": "number"
          }
        },
        "required": [
          "planting_id",
          "bed_id",
          "taxon_id",
          "name",
          "wants",
          "gets",
          "sun_hours",
          "problem"
        ],
        "title": "MisplacedOut",
        "type": "object"
      },
...
      "ShadowDay": {
        "properties": {
          "day": {
            "title": "Day",
            "type": "integer"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ShadowFrame"
            },
            "title": "Frames",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "day",
          "frames"
        ],
        "title": "ShadowDay",
        "type": "object"
      },
      "ShadowFrame": {
        "description": "Every shadow in the garden at one moment of one day.",
        "properties": {
          "altitude": {
            "title": "Altitude",
            "type": "number"
          },
          "azimuth": {
            "title": "Azimuth",
            "type": "number"
          },
          "minute": {
            "title": "Minute",
            "type": "integer"
          },
          "polygons": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "title": "Polygons",
            "type": "array"
          }
        },
        "required": [
          "minute",
          "altitude",
          "azimuth",
          "polygons"
        ],
        "title": "ShadowFrame",
        "type": "object"
      },
...
      "TerrainOut": {
        "description": "The ground under a garden, and how far it is to be trusted.\n\nEvery field after `relief` is there so the page can answer \"says who, and\nhow good is it\" without the reader having to know what a DGM1 is. A height\nshown without its credit is a height used outside its licence, and a height\nshown without its accuracy invites more confidence than it earns.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "cell_m": {
            "title": "Cell M",
            "type": "number"
          },
          "cols": {
            "title": "Cols",
            "type": "integer"
          },
          "highest": {
            "title": "Highest",
            "type": "number"
          },
          "licence": {
            "title": "Licence",
            "type": "string"
          },
          "lowest": {
            "title": "Lowest",
            "type": "number"
          },
          "min_x": {
            "title": "Min X",
            "type": "number"
          },
          "min_y": {
            "title": "Min Y",
            "type": "number"
          },
          "relief": {
            "items": {
              "type": "number"
            },
            "title": "Relief",
            "type": "array"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "vertical_step_m": {
            "title": "Vertical Step M",
            "type": "number"
          }
        },
        "required": [
          "cell_m",
          "min_x",
          "min_y",
          "cols",
          "rows",
          "relief",
          "lowest",
          "highest",
          "source",
          "licence",
          "attribution",
          "vertical_step_m"
        ],
        "title": "TerrainOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NinaNatur",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/gardens": {
      "post": {
        "description": "Create a garden. The token comes back here and is the only way back in.\n\nA signed-in visitor's garden is theirs from the moment it exists. It used to\nbe nobody's unless something called `/claim` afterwards, and nothing did \u2014\nso an account kept the links and then listed none of them.\n\n`current_account` rather than `require_account`: making a garden without an\naccount is the ordinary case, and the token is still the whole of its\naccess control either way.",
        "operationId": "create_api_v1_gardens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GardenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GardenCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "gardens"
        ]
      }
    },
...
    "/api/v1/gardens/{token}/light": {
      "get": {
        "description": "The stored map, or null when nothing has been drawn yet.\n\n`month` asks for one month instead of the season, computed on the spot and\nnot stored. March to October, the same window the whole light model uses:\na plant's December is not what decides where it can live, and a map of it\nwould drag every German garden into shade.\n\nA month is a computation \u2014 about a quarter of a relight, seconds on a big\ngarden \u2014 so it takes a slot and counts against the visitor's limit like the\nother three, slot first. The stored season map is a read, never turned away.",
        "operationId": "light_map_api_v1_gardens__token__light_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10,
                  "minimum": 3,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LightMap"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Light Map Api V1 Gardens  Token  Light Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Light Map",
        "tags": [
          "light"
        ]
      },
      "post": {
        "description": "Recompute the whole map, now, because somebody asked.\n\nBelt as well as braces. The signature should catch every change that moves a\nshadow, and if it ever does not, this is how somebody fixes their own map\nwithout knowing why it was wrong.\n\nIt is also where a garden gets its ground for the first time. A state survey\ntakes seconds to answer, which is too long for a page load and perfectly\nreasonable for a button \u2014 and afterwards every recompute reads it for free.",
        "operationId": "rebuild_light_map_api_v1_gardens__token__light_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LightMap"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Rebuild Light Map Api V1 Gardens  Token  Light Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Light Map",
        "tags": [
          "light"
        ]
      }
    },
...
    "/api/v1/gardens/{token}/shadows": {
      "get": {
        "description": "Where the shadows fall through one middling day of a month.\n\nThe 15th, because a month's first and last days differ by a fortnight of sun\nand the middle is the one that represents it. Computed rather than stored:\nit is one day rather than a season, and nobody watches it twice in a row.",
        "operationId": "shadows_through_a_day_api_v1_gardens__token__shadows_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 12,
              "minimum": 1,
              "title": "Month",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowDay"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shadows Through A Day",
        "tags": [
          "light"
        ]
      }
    },
...
    "/api/v1/gardens/{token}/sources": {
      "get": {
        "description": "Every survey this garden's numbers actually rest on (doc 106).\n\nEmpty is an ordinary answer: a garden nobody has measured anything for owes\nnobody anything. What is here is what was used, not what the state could in\nprinciple offer \u2014 a credit for a source a garden never touched would be a\nclaim about where its numbers came from.",
        "operationId": "sources_api_v1_gardens__token__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CreditOut"
                  },
                  "title": "Response Sources Api V1 Gardens  Token  Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sources",
        "tags": [
          "light"
        ]
      }
    },
    "/api/v1/gardens/{token}/terrain": {
      "get": {
        "description": "The ground under this garden, or null where nobody publishes it.\n\nNull is an answer and the page says so in words. Nine Bundesl\u00e4nder have no\nservice in the registry, and a garden there keeps the flat assumption it\nalways had \u2014 which is fine, and being quiet about it is not.",
        "operationId": "terrain_api_v1_gardens__token__terrain_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TerrainOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Terrain Api V1 Gardens  Token  Terrain Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Terrain",
        "tags": [
          "light"
        ]
      }
    },
...
  }
}